chromosome map parser -> chromosome component
