+++ projects/utils/pyproject.toml
@@ -24,7 +24,6 @@ requires-python = ">=3.11,<4.0"
 
 dependencies = [
     "python-dateutil (>=2.9.0.post0,<3.0.0)",
-    "pandas (>=2.2.3,<3.0.0)",
     "strenum (>=0.4.15,<0.5.0)",
 ]
