+++ test/components/niagads/chromosome_map_parser/test_core.py
@@ -1,0 +1,5 @@
+from niagads.chromosome_map_parser import core
+
+
+def test_sample():
+    assert core is not None
